backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
httpx==0.28.1
dnspython==2.7.0
python-magic==0.4.27
beautifulsoup4==4.12.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
yara-python==4.5.1
pypdf==5.1.0
